Added `services` directory in a`shared*` packages
