+++ .prettierrc
@@ -10,6 +10,8 @@
     "",
     "<THIRD_PARTY_MODULES>",
     "",
+    "^(@shared)(/services/.*)$",
+    "",
     "^(@shared)(/lib/.*)$",
     "",
     "^(@shared)(/types/.*)$",
@@ -20,11 +22,15 @@
     "",
     "^(@sharedClient)(/stores/.*)$",
     "",
+    "^(@sharedClient)(/services/.*)$",
+    "",
     "^(@sharedClient)(/lib/.*)$",
     "",
     "^(@sharedClient)(/hooks/.*)$",
     "",
     "^(@sharedClient)(/.*)$",
+    "",
+    "^(@sharedServer)(/services/.*)$",
     "",
     "^(@sharedServer)(/lib/.*)$",
     "",
