+++ eslint.config.mjs
@@ -3,7 +3,7 @@ import reactHooks from 'eslint-plugin-react-hooks';
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
 
-const sharedLanguageOptions = {
+const SHARED_LANGUAGE_OPTIONS = {
   ecmaVersion: 2022,
   sourceType: 'module',
   parser: tseslint.parser,
@@ -13,26 +13,38 @@ const sharedLanguageOptions = {
   },
 };
 
-const noRelativeImportsPattern = {
+const NO_RELATIVE_IMPORTS_PATTERN = {
   group: ['.*'],
   message: 'Use imports like `@src` and `@shared` instead of relative paths.',
 };
 
-const noFirebaseAdminImportPattern = {
+const NO_FIREBASE_ADMIN_IMPORT_PATTERN = {
   group: ['firebase-admin', 'firebase-admin/*'],
   message:
     'Importing from the `firebase-admin` library is not allowed in this package. Use the `firebase` client-side library instead.',
 };
 
-const noFirebaseClientImportPattern = {
+const NO_FIREBASE_CLIENT_IMPORT_PATTERN = {
   group: ['firebase', 'firebase/*'],
   message:
     'Importing from the client-side `firebase` library is not allowed in this package. Use the `firebase-admin` library instead.',
 };
 
+const NO_SHARED_CLIENT_IMPORT_PATTERN = {
+  group: ['@sharedClient', '@sharedClient/*'],
+  message: 'Importing from the `@sharedClient` package is not allowed in this package.',
+};
+
+const NO_SHARED_SERVER_IMPORT_PATTERN = {
+  group: ['@sharedServer', '@sharedServer/*'],
+  message: 'Importing from the `@sharedServer` package is not allowed in this package.',
+};
+
 function makeSharedRules({
   disallowFirebaseAdminImports = false,
   disallowFirebaseClientImports = false,
+  disallowSharedClientImports = false,
+  disallowSharedServerImports = false,
 }) {
   return {
     'no-console': 'error',
@@ -52,9 +64,11 @@ function makeSharedRules({
       'error',
       {
         patterns: [
-          noRelativeImportsPattern,
-          disallowFirebaseAdminImports ? noFirebaseAdminImportPattern : null,
-          disallowFirebaseClientImports ? noFirebaseClientImportPattern : null,
+          NO_RELATIVE_IMPORTS_PATTERN,
+          disallowFirebaseAdminImports ? NO_FIREBASE_ADMIN_IMPORT_PATTERN : null,
+          disallowFirebaseClientImports ? NO_FIREBASE_CLIENT_IMPORT_PATTERN : null,
+          disallowSharedClientImports ? NO_SHARED_CLIENT_IMPORT_PATTERN : null,
+          disallowSharedServerImports ? NO_SHARED_SERVER_IMPORT_PATTERN : null,
         ].filter((p) => p !== null),
       },
     ],
@@ -89,15 +103,19 @@ export default tseslint.config(
   // Shared models/lib package config.
   {
     files: ['packages/shared/src/**/*.ts'],
-    languageOptions: sharedLanguageOptions,
-    rules: makeSharedRules({disallowFirebaseAdminImports: true}),
+    languageOptions: SHARED_LANGUAGE_OPTIONS,
+    rules: makeSharedRules({
+      disallowFirebaseAdminImports: true,
+      disallowSharedClientImports: true,
+      disallowSharedServerImports: true,
+    }),
   },
 
   // Shared client package config.
   {
     files: ['packages/sharedClient/**/*.{ts,tsx}'],
     languageOptions: {
-      ...sharedLanguageOptions,
+      ...SHARED_LANGUAGE_OPTIONS,
       globals: globals.browser,
     },
     plugins: {
@@ -105,7 +123,10 @@ export default tseslint.config(
     },
     rules: {
       ...reactHooks.configs.recommended.rules,
-      ...makeSharedRules({disallowFirebaseAdminImports: true}),
+      ...makeSharedRules({
+        disallowFirebaseAdminImports: true,
+        disallowSharedServerImports: true,
+      }),
     },
   },
 
@@ -113,15 +134,18 @@ export default tseslint.config(
   // TODO: Figure out why this is not causing lint errors for `type` imports.
   {
     files: ['packages/sharedServer/**/*.{ts}'],
-    languageOptions: sharedLanguageOptions,
-    rules: makeSharedRules({disallowFirebaseClientImports: true}),
+    languageOptions: SHARED_LANGUAGE_OPTIONS,
+    rules: makeSharedRules({
+      disallowFirebaseClientImports: true,
+      disallowSharedClientImports: true,
+    }),
   },
 
   // PWA package config.
   {
     files: ['packages/pwa/**/*.{ts,tsx}'],
     languageOptions: {
-      ...sharedLanguageOptions,
+      ...SHARED_LANGUAGE_OPTIONS,
       globals: globals.browser,
     },
     plugins: {
@@ -129,16 +153,22 @@ export default tseslint.config(
     },
     rules: {
       ...reactHooks.configs.recommended.rules,
-      ...makeSharedRules({disallowFirebaseAdminImports: true}),
+      ...makeSharedRules({
+        disallowFirebaseAdminImports: true,
+        disallowSharedServerImports: true,
+      }),
     },
   },
 
   // Scripts package config.
   {
     files: ['packages/scripts/**/*.ts'],
-    languageOptions: sharedLanguageOptions,
+    languageOptions: SHARED_LANGUAGE_OPTIONS,
     rules: {
-      ...makeSharedRules({disallowFirebaseClientImports: true}),
+      ...makeSharedRules({
+        disallowFirebaseClientImports: true,
+        disallowSharedClientImports: true,
+      }),
       // TODO: Remove this after getting @src imports working in /scripts.
       'no-restricted-imports': 'off',
     },
@@ -148,7 +178,7 @@ export default tseslint.config(
   {
     files: ['packages/extension/**/*.{ts,tsx}'],
     languageOptions: {
-      ...sharedLanguageOptions,
+      ...SHARED_LANGUAGE_OPTIONS,
       globals: globals.browser,
     },
     plugins: {
@@ -156,14 +186,20 @@ export default tseslint.config(
     },
     rules: {
       ...reactHooks.configs.recommended.rules,
-      ...makeSharedRules({disallowFirebaseAdminImports: true}),
+      ...makeSharedRules({
+        disallowFirebaseAdminImports: true,
+        disallowSharedServerImports: true,
+      }),
     },
   },
 
   // Functions package config.
   {
     files: ['packages/functions/**/*.{ts}'],
-    languageOptions: sharedLanguageOptions,
-    rules: makeSharedRules({disallowFirebaseClientImports: true}),
+    languageOptions: SHARED_LANGUAGE_OPTIONS,
+    rules: makeSharedRules({
+      disallowFirebaseClientImports: true,
+      disallowSharedClientImports: true,
+    }),
   }
 );
