+++ packages/pwa/src/components/auth/SignOutRedirect.tsx
@@ -1,7 +1,8 @@
 import {useEffect} from 'react';
 import {useNavigate} from 'react-router-dom';
 
-import {logger} from '@shared/lib/logger';
+import {logger} from '@shared/services/logger';
+
 import {Urls} from '@shared/lib/urls';
 
 import {authService} from '@sharedClient/lib/auth.client';
