+++ packages/pwa/src/lib/shortcuts.pwa.ts
@@ -1,5 +1,5 @@
-import {KeyboardShortcutsService} from '@shared/lib/shortcuts';
+import {SharedKeyboardShortcutsService} from '@shared/services/keyboardShortcuts';
 
 import {IS_MAC} from '@src/lib/environment.pwa';
 
-export const keyboardShortcutsService = new KeyboardShortcutsService({isMac: IS_MAC});
+export const keyboardShortcutsService = new SharedKeyboardShortcutsService({isMac: IS_MAC});
