+++ packages/sharedClient/src/lib/auth.client.ts
@@ -1,5 +1,5 @@
-import {AuthService} from '@shared/lib/auth';
+import {SharedAuthService} from '@shared/services/auth';
 
 import {firebaseService} from '@sharedClient/lib/firebase.client';
 
-export const authService = new AuthService(firebaseService.auth);
+export const authService = new SharedAuthService(firebaseService.auth);
